[FAB-17133] Add branch to AZP trigger
Add the branch name, i.e., refs/pull/3/merge to
the AZP trigger. Without the branch it attempts
to checkout the commit which does not exist in
the repo, as it comes from the remote branch of
the user. This makes the checkout use
refs/remote/pull/3/merge which resolves correctly
to the users repo.

Signed-off-by: Brett Logan <Brett.T.Logan@ibm.com>